Moving hdfs_connect to hdfs module
